+++ Documentation/Changelog.md
@@ -7,6 +7,7 @@
 Release date: not released yet
 
 - `Unit` type now implements `IEquatable<Unit>` and `IComparable<Unit>` interfaces
+- Added `GetFileSize(string)` method to `IReadOnlyFileSystemView`
 - `SemanticVersion` type can be parsed by `System.Text.Json` serializer out of the box (.NET 8.0+)
 - `Interval<T>` and `ValueInterval<T>` types can be parsed by `System.Text.Json` serializer out of the box (.NET 8.0+)
 - Added type converters for `Interval<T>` and `ValueInterval<T>` types